+++ devel/CHANGELOG.md
@@ -15,7 +15,11 @@
 - [typo]: typography;
 - [warn]: fixing compiler warning messages;
 
-# #2025-04-19
+## 2026-02-11
+
+- TOL: addinf 'UndefinedTimeStamp';
+
+## 2025-04-19
 
 - CSDMNC: [adpt];
 - XDHCDC: the parameter given as id for a connect action is no more only a user id, but contains more parameters;